pyproject.toml
[project]
name = "stackit_cost_monitoring"
version = "0.1.0"
description = ""
authors = [
    {name = "Harald Wilhelmi",email = "harald.wilhelmi@uni-heidelberg.de"}
]
license = {text = "GNU Public License v3"}
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "pydantic",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry.group.dev.dependencies]
openapi-generator-cli = "^7.14.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
check_stackit_costs = "stackit_cost_monitoring.nagios_plugin:main"
